dj_prj/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.10
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.3
django-baton==4.2.2
django-cachalot==2.7.0
django-filter==24.3
django-localflavor==4.0
django-redis==5.4.0
django-versatileimagefield==3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
flower==2.0.1
humanize==4.11.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
python-stdnum==1.20
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rollbar==1.1.2
rpds-py==0.21.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tornado==6.4.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
